use signals to coordinate setting up of the logger between logger.py and executor.py
